currencies dialog: rates instead of worths; a worth is i64
A currency read "worth 1 175 000", a number that only meant something beside the others,
and the dialog kept asking to multiply them all. Now a row reads "1 EUR = 117,5 RSD",
against the cheapest currency of the tour, and the scale behind it is the dialog's own
business: raised quietly when a typed rate or today's rate needs it (rates::room). The
offer to multiply and the two-zeros rule are gone.

A worth is i64, not the C#'s int: nothing reads the database in C# any more, and the
scale no longer has to be squeezed under 2 147 483 647. Old tours read as they were.

+++ rust/crates/tc-core/src/domain.rs
@@ -34,7 +34,7 @@ pub struct Currency {
     pub id: CurrencyId,
     pub name: String,
     /// Rate against the tour's base, times 100 - as it is stored today.
-    pub rate: i32,
+    pub rate: i64,
     pub extras: Extras,
 }
 
@@ -548,7 +548,7 @@ pub mod wire {
         #[serde(default = "coin")]
         pub name: String,
         #[serde(default = "hundred")]
-        pub currency_rate: i32,
+        pub currency_rate: i64,
         /// Everything this struct did not name. See [`Extras`].
         #[serde(flatten)]
         pub rest: serde_json::Map<String, serde_json::Value>,
@@ -557,7 +557,7 @@ pub mod wire {
     fn coin() -> String {
         "coin".to_owned()
     }
-    fn hundred() -> i32 {
+    fn hundred() -> i64 {
         100
     }
